requirements.txt
# Docling OCR Server Requirements
# ================================
# Install with: pip install -r requirements.txt

# FastAPI Server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Docling OCR
docling>=2.0.0

# PDF Processing
pdf2image>=1.16.0
pypdfium2>=4.0.0

# Image Processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Table Handling
pandas>=2.0.0
openpyxl>=3.1.0

# Text Correction
wordninja>=2.0.0

# HTTP Requests (for testing)
requests>=2.28.0
